+++ react/bl-react-app/src/App.js
@@ -1,61 +1,14 @@
 import React from 'react';
-import { MemoryRouter, Switch, Route } from 'react-router-dom';
-
-import Container from 'react-bootstrap/Container';
-import Button from 'react-bootstrap/Button';
-import ButtonToolbar from 'react-bootstrap/ButtonToolbar';
-import { LinkContainer } from 'react-router-bootstrap';
 
 import './css/w3.css';
 import './css/App.css';
-import About from './Components/About.js';
-import W3NavBar1 from './Components/w3/NavBar1.js';
+import AppHome from './Components/Home.js';
+import W3Blog from './Components/w3/Blog.js';
 
-const Home = () => <span>Home</span>;
- 
 
-const Users = () => <span>Users</span>;
 
 const App = () => (
-  <MemoryRouter>
-    <Container className="p-3">
-      <Container className="p-5 mb-4 bg-light rounded-3">
-        <h1 className="header">Welcome To React-Bootstrap v0.54</h1>
-        <h2>
-          <W3NavBar1 />
-          Navigate to{' '}
-          <ButtonToolbar className="custom-btn-toolbar">
-            <LinkContainer to="/">
-              <Button>Home</Button>
-            </LinkContainer>
-            <LinkContainer to="/about">
-              <Button>About</Button>
-            </LinkContainer>
-            <LinkContainer to="/users">
-              <Button>Users</Button>
-            </LinkContainer>
-          </ButtonToolbar>
-        </h2>
-        <h2>
-          Current Page is{' '}
-          <Switch>
-            <Route path="/about">
-              <About />
-            </Route>
-            <Route path="/users">
-              <Users />
-            </Route>
-            <Route path="/">
-              <Home />
-            </Route>
-          </Switch>
-        </h2>
-        
-      </Container>
-
-    </Container>
-    
-  </MemoryRouter>
+  <W3Blog />
 );
 
 export default App;
