+++ react/bl-react-app/src/Components/About.js
@@ -5,7 +5,7 @@ import MainApp from '../Main/MainApp.js';
 const About = () => (
     
     <Container>
-    <span>About:xdv0.14</span>    
+    <span>About:xdv0.15</span>    
     <div className="App">
       <MainApp />
     </div>
