+++ react/bl-react-app/src/Components/w3/Blog.js
@@ -1,0 +1,16 @@
+
+import Container from 'react-bootstrap/Container'; 
+
+import '../../css/w3.css';
+const W3Blog = () => (    
+    <Container>
+    <div class="w3-bar w3-black">
+        <a href=".." class="w3-bar-item w3-button">W3Blog_xdv0.13</a>
+        <a href="http://baidu.com" class="w3-bar-item w3-button">baidu</a>
+        <a target="_blank" href="http://localhost:8080" class="w3-bar-item w3-button">8080</a>
+        <a target="_blank" href="http://localhost:3000" class="w3-bar-item w3-button">3000</a>
+    </div> 
+    </Container>
+);
+
+export default W3Blog;
